refactor(sip): extract the DTMF feature codes into DtmfFeatureCodes
Move onDtmfInfo (Signal= parsing, the per-Call-ID DtmfAccum accumulator,
the *60/*80/*73/*69/*11/*72 CLASS codes, and the *4887 / *PIN#code admin
menu including the Issue #93 star-code quirk) plus _adminExt and its NVS
load/save out of RequestsHandler into DtmfFeatureCodes. endCall()'s
accumulator erase becomes _dtmf.forgetCall(callId). tick()'s sweep becomes
_dtmf.sweepStale(), which tests _env.findSession(callId) instead of the
engine's _sessions map -- equivalent because every _sessions emplace is
guarded and the map never holds a null shared_ptr.

DtmfFeatureCodes takes PbxEnv& (log/enqueue/findRegistered/findSession/
validAor/localIp -- all pre-existing virtuals), PbxFeatureConfig& and
CdrRing& by reference, and a std::function<uint16_t()> wired to
grantAdminHttpGraceWindow(). The *4887 branch used to inline the same
epoch-ms + TTL arithmetic that method already does; calling back avoids
a third copy of the steady_clock read. grantAdminHttpGraceWindow() now
returns the TTL it applied (was void) so the DTMF log line can report it
-- the one public-signature change in this decomposition; HttpServer's
existing caller ignores the return. The DTMF path now loads
_adminHttpTtlSec once instead of twice, so the stored deadline and the
logged TTL can no longer disagree under a concurrent write.

Scope renames in the moved body: queueLog->_env.log; findClient->
_env.findRegistered and getSession->_env.findSession (optional<shared_ptr>
-> bare shared_ptr, so .has_value()/.value()-> become if/->);
isValidAor->_env.validAor; _localIp->_env.localIp(); _outbox.emplace_back->
_env.enqueue; getMessageFromPool->sipmsgpool::getMessageFromPool.
saveAdminExt has no callers and moves as-is. RequestsHandler keeps
getAdminExt() as a one-line forwarder (tests and HttpServer call it).

Host 309/309 (excl. one pre-existing flaky test), firmware compiles for
SIP_TRANSPORT=eth, callgraph guard clean. Adversarial verify: six lenses,
all clean.
