+++ main/CMakeLists.txt
@@ -93,6 +93,7 @@ idf_component_register(SRCS ${MAIN_SRC}
                             "../src/SIP/SipMessagePool.cpp"
                             "../src/SIP/PbxFeatureConfig.cpp"
                             "../src/SIP/CdrRing.cpp"
+                            "../src/SIP/DtmfFeatureCodes.cpp"
                             "../src/SIP/RequestsHandler.cpp"
                             "../src/SIP/TransactionLayer.cpp"
                             "../src/SIP/Registrar.cpp"
